README.rst: Drop recommendation to install meson using pip3
Usage of pip3 to install meson may result in a different meson version
available for the local user and for root. This causes issues when
running `meson install` (directly or through `ninja install`):
Installing to a system location will request elevated privileges and use
the system meson version to install files, which may fail to unpickle
data pickled with the meson version for the local user.

As all major distributions now provide a recent-enough version of meson,
usage of pip3 is not longer needed in most cases, and shouldn't be
recommended. Drop it.

While at it, update the version of meson listed in the documentation to
match the required version.

Bug: https://bugs.libcamera.org/show_bug.cgi?id=199
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>