pipeline: rpi: pisp: Add DMA syncronisation
When an application allocates it's own DMA buffers for the pipeline
output, any CPU access to these buffers needs to be bracketed by
DMA_SYNC ioctls. This is
to prevent any cache coherency issues.

When dequeuing the  output buffers, the pipeline may perform software
downscaling and format conversions on the output buffers, so this access
must be bracketed with a read/write sync to allow an application to use
DMA buffers.

Signed-off-by: William Vinnicombe <william.vinnicombe@raspberrypi.com>