libcamera: delayed_controls: Operate on a CameraSensor
The DelayedControls class currently operates on a V4L2Device, which it
takes as an argument to the constructor. Controls are accessed on the
device through its controls(), getControls() and setControls()
functions. This makes the DelayedControls V4L2-specific, and gets in the
way of moving the CameraSensor class to expose libcamera controls
instead of V4L2 controls.

In preparation for the CameraSensor controls rework, modify
DelayedControls to operate on a CameraSensor instead of a V4L2Device.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>